workers/geo-pipeline/requirements.txt
rasterio>=1.3.0
geopandas>=0.14.0
numpy>=1.26.0
pymongo>=4.6.0
requests>=2.31.0
shapely>=2.0.0
structlog>=24.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
